requirements.txt
asgiref==3.2.3
boto3==1.12.33
botocore==1.15.33
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-nine==0.2.3
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
gunicorn==20.0.4
jmespath==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
